Tighten handling around non-numeric PKs
Validate the type of the primary key against a list of postgres numeric
types. Previously this logic relied on catching an undefined function
error due to application of "max" to a non-numeric type. This worked for
UUID, but not for text, since text supports max.

The resulting behavior was a confusing "nothing to fill" error
downstream.