Sort the partitions before using them (#31)
I had an issue where the `fill` command turned out to not insert anything as it was relying on the order of `Table#partitions` to be sorted with the oldest partition first, and newest partition last. Mine was returning partitions with the newest first, oldest last, which generated fill statements with out of order dates (`starting_time` was the newest partition, and `ending_time` was the oldest partition).

Example:

```sql
... WHERE "id" > X AND "id" <= Y AND "time" >= '2020-06-01'::date AND "time" < '2020-03-31'::date;
```

This change does rely on the partitions being lexicographical sortable for the purposes of date though.