Minor readme suggestions
Mention range partitioning

There are additional partitioning types like list partitioning. If
pgslice only supports range partitioning, for those familiar with the
types, it may be worth mentioning the type. It helps explain why the
recommendation to pick a column would be for a timestamp column.

Show the output of `SHOW constraint_exclusion;` to help clarify